feat(org): move worker-runtime config from helix-org settings to Providers
Extract the worker.* dropdowns (runtime/credentials/provider/model) into a
shared WorkerRuntimePanel and render it on the AI Providers page, next to the
providers it references. Strip them from HelixOrgSettings (worker keys stay in
FIRST_CLASS_KEYS so they don't reappear as generic rows). Alpha-gated on
helix-org and org context.
